feat(shared): ブックマーク一覧取得 (GET_BOOKMARKS) のスキーマ定義（テスト定数化の徹底） (#444)
